requirements.txt
aiogram==2.14.3
aiogram-dialog==1.0.1
aiohttp==3.7.4.post0
aioredis==1.3.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
hiredis==2.0.0
idna==3.2
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
